rename schema.py, fix pep8 errors
