+++ scraper/WikiFetcher.py
@@ -1,16 +1,15 @@
-import requests
-import unittest
+from .schema import schema_map, schema
+from .urls import parliaments
+import logging
 import pandas as pd
 import os
 from datetime import datetime
 import ssl
 ssl._create_default_https_context = ssl._create_unverified_context
 
-import logging
+
 logging.basicConfig(level=logging.INFO, format='%(asctime)s %(message)s')
 
-from .urls import parliaments
-from .schema_map import schema_map, schema
 
 class Wikifetcher:
 
@@ -53,8 +52,11 @@ class Wikifetcher:
         for parliament, parliament_data in parliaments.items():
             logging.info("Fetching parliament of %s" % parliament_data["name"])
             politicians_table, table_index = self.get_politicians(parliament_data["url"])
-            logging.info("Retrieved %d entries from table %d" % (politicians_table.shape[0], table_index))
-            politicians_table.to_csv(output_path + strpdatetoday + "_" + parliament + ".csv", index = False, header = True)
+            logging.info("Retrieved %d entries from table %d" %
+                         (politicians_table.shape[0], table_index))
+            politicians_table.to_csv(
+                output_path + strpdatetoday + "_" + parliament + ".csv", index=False, header=True)
+
 
 if __name__ == "__main__":
     fetcher = Wikifetcher()
