+++ scraper/schema.py
